- Added receipt uploading feature (needs refactoring) - cleaned up property injection - formated all files - cleaned up unused imports

+++ src/main/java/nl/brandfonds/Brandfonds/resource/AuthenticationController.java
@@ -16,10 +16,13 @@ import nl.brandfonds.Brandfonds.model.PasswordChangeRequest;
 import nl.brandfonds.Brandfonds.model.RegisterRequest;
 import nl.brandfonds.Brandfonds.model.User;
 import nl.brandfonds.Brandfonds.security.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.authentication.DisabledException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
@@ -54,7 +57,7 @@ public class AuthenticationController {
     AuthenticationManager authenticationManager;
 
     @PostMapping
-    @ApiOperation(value = "Login", notes = "Logs in user using forname and password")
+    @ApiOperation(value = "Login", notes = "Logs in user using forname and password", nickname = "login")
     @ApiResponses({
             @ApiResponse(code = 200, message = "User was successfully logged in", response = AuthenticationResponse.class),
             @ApiResponse(code = 403, message = "The user is set to inactive", response = ResponseEntity.class),
@@ -67,13 +70,11 @@ public class AuthenticationController {
                     authenticationRequest.getPassword()));
         } catch (BadCredentialsException e) {
             throw new NotFoundException("De ingevoerde voornaam of het wachtwoord is fout");
+        } catch (DisabledException d){
+            throw new UserDisabledException("Het account waarmee je probeert in te loggen is uitgeschakeld, wacht of neem contact op met de brandmeester totdat je account is geactiveerd");
         }
 
         final CustomUserDetails userDetails = (CustomUserDetails) customUserDetailsService.loadUserByUsername(authenticationRequest.getUsername());
-
-        if (!userDetails.isEnabled()) {
-            throw new UserDisabledException("Het account waarmee je probeert in te loggen is uitgeschakeld, wacht of neem contact op met de brandmeester totdat je account is geactiveerd");
-        }
 
         final String jwt = jwtTokenUtil.generateToken(userDetails);
 
@@ -84,7 +85,7 @@ public class AuthenticationController {
     //region Register methods
 
     @PostMapping(path = "/register")
-    @ApiOperation(value = "Register", notes = "Creates a register request for user and sends verification mail", response = ResponseEntity.class)
+    @ApiOperation(value = "Register", notes = "Creates a register request for user and sends verification mail", nickname = "register")
     @ApiResponses({
             @ApiResponse(code = 200, message = "The registration request was successfully created", response = ResponseEntity.class),
             @ApiResponse(code = 409, message = "There is already a user registered with the same mail adres", response = ResponseEntity.class),
@@ -103,12 +104,16 @@ public class AuthenticationController {
 
 
     @GetMapping(path = "/registerconformation/{randomstring}")
-    @ApiOperation(value = "Confirm Registration", notes = "Validates registration with generated string and sends activation request to brandmaster")
+    @ApiOperation(value = "Confirm Registration", notes = "Validates registration with generated string and sends activation request to brandmaster", nickname = "registerConformation")
     @ApiResponses({
             @ApiResponse(code = 200, message = "The register request was successfully validated", response = ResponseEntity.class)
             //todo add response for link expired -> add correct code to exception class
     })
     public void confirmRegistration(@PathVariable("randomstring") String randomstring) throws LinkExpiredException {
+
+        if (!registerRequestService.getByrandomString(randomstring).isPresent()) {
+            throw new LinkExpiredException("De gebruikte link is ongeldig of verlopen.");
+        }
 
         registerRequestService.getByrandomString(randomstring).ifPresent(request -> {
 
@@ -120,11 +125,7 @@ public class AuthenticationController {
                     request.getEmailadres(), request.getForname() + " "
                             + request.getSurname(),
                     (userService.getByMail(request.getEmailadres()).get().getId()));
-
-            return;
         });
-
-        throw new LinkExpiredException("De gebruikte link is ongeldig of verlopen.");
     }
 
     //endregion
@@ -132,74 +133,79 @@ public class AuthenticationController {
     //region Edit passwords methods
 
     @GetMapping(path = "/forgotpassword/{mailadres}")
-    @ApiOperation(value = "Forgot password", notes = "Checks if user exist and sends password change mail to corresponding mail address")
+    @ApiOperation(value = "Forgot password", notes = "Checks if user exist and sends password change mail to corresponding mail address", nickname = "resetPasswordRequest")
     @ApiResponses({
             @ApiResponse(code = 200, message = "Mail was successfully send to corresponding address", response = ResponseEntity.class),
             @ApiResponse(code = 404, message = "The provided emailadress is not found", response = ResponseEntity.class)
     })
     public void forgotPassword(@PathVariable("mailadres") String mailadres) throws NotFoundException {
 
+        if (!userService.getByMail(mailadres).isPresent()){
+            throw new NotFoundException("Er kan geen gebruiker met dit mailadres worden gevonden");
+        }
+
         userService.getByMail(mailadres).ifPresent(user -> {
                     PasswordChangeRequest request = new PasswordChangeRequest(mailadres);
                     passwordChangeRequestService.save(request);
 
                     mailService.sendChangePasswordMail(request.getEmailadres(), request.getRandomstring());
-                    return;
                 }
         );
-
-        throw new NotFoundException("Er kan geen gebruiker met dit mailadres worden gevonden");
     }
 
 
     @GetMapping(path = "/resetpasswordcode/{randomstring}")
-    @ApiOperation(value = "Forgot password validation", notes = "Validated if given string is linked to password change request")
+    @ApiOperation(value = "Forgot password validation", notes = "Validated if given string is linked to password change request", nickname = "validateLinkPasswordRequest")
     @ApiResponses({
             @ApiResponse(code = 200, message = "Change password string is validated", response = ResponseEntity.class),
             @ApiResponse(code = 401, message = "The token provided is expired/ not found", response = ResponseEntity.class)
             //todo add response for link expired -> add correct code to exception class
     })
-    public void checkPasswordLink(@PathVariable("randomstring") String randomstring) throws LinkExpiredException {
+    public void checkPasswordResetLink(@PathVariable("randomstring") String randomstring) throws LinkExpiredException {
         if (!passwordChangeRequestService.getByrandomString(randomstring).isPresent()) {
-            throw new LinkExpiredException("De link die je probeert te gebruiken is verlopen");
+            throw new LinkExpiredException("De link die je probeert te gebruiken bestaat niet of is verlopen");
         }
     }
 
 
     @PostMapping(path = "/resetpassword/{randomstring}")
-    @ApiOperation(value = "Change password", notes = "Changes password to new value")
+    @ApiOperation(value = "Change password", notes = "Changes password to new value", nickname = "confirmPasswordRequest")
     @ApiResponses({
             @ApiResponse(code = 200, message = "Password was succesfully changed", response = ResponseEntity.class),
             @ApiResponse(code = 401, message = "The token provided is expired/ not found", response = ResponseEntity.class)
     })
     public void changePassword(@PathVariable("randomstring") String randomstring, @RequestBody String password) throws LinkExpiredException {
+
+        if (!passwordChangeRequestService.getByrandomString(randomstring).isPresent()) {
+            throw new LinkExpiredException("De link die je probeert te gebruiken bestaat niet of is verlopen");
+        }
+
         passwordChangeRequestService.getByrandomString(randomstring).ifPresent(request -> {
             userService.updatePassword(passwordEncoder.encode(password), request.getEmailadres());
             passwordChangeRequestService.delete(request);
         });
-        throw new LinkExpiredException("De link die je probeert te gebruiken is verlopen");
     }
     //endregion
 
 
     @GetMapping(path = "/activate-user/{id}/{is-activated}")
-    @ApiOperation(value = "Set user activation", notes = "Set user active/inactive")
+    @ApiOperation(value = "Set user activation", notes = "Set user active/inactive", nickname = "setUserActivation")
     @ApiResponses({
             @ApiResponse(code = 200, message = "User status successfully updated", response = ResponseEntity.class),
             @ApiResponse(code = 404, message = "User could not be found", response = ResponseEntity.class)
     })
-    public void activateUser(@PathVariable("id") Integer id, @PathVariable("is-activated") boolean isActivated) throws UserDisabledException {
+    public void setUserActivation(@PathVariable("id") Integer id, @PathVariable("is-activated") boolean isActivated) throws UserDisabledException {
+
+        if (!userService.getByID(id).isPresent()) {
+            throw new UserDisabledException("De gebruiker die je wilt activeren staat niet meer in het systeem");
+        }
 
         userService.getByID(id).ifPresent(user -> {
             user.setActivated(isActivated);
             userService.save(user);
 
             this.mailService.sendUserActivatedMail(user.getEmailadres(), user.getEmailadres(), user.getForname());
-            return;
         });
-
-        throw new UserDisabledException("De gebruiker die je wilt activeren staat niet meer in het systeem");
-
     }
 
 }
